requirements.txt
appnope==0.1.0
asgiref==3.2.7
backcall==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
coverage==4.0.3
decorator==4.4.2
Django==3.0.5
django-cors-middleware==1.5.0
django-environ==0.4.5
django-filter==2.2.0
django-money==1.1
django-oauth-toolkit==1.3.2
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==4.0.3
freezegun==0.3.15
idna==2.9
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
oauthlib==3.1.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
py-moneyed==0.8.0
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
soft-delete==0.2.2
sqlparse==0.3.1
text-unidecode==1.3
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.9
